requirements.txt
onnxruntime-gpu
museval
resampy